Refactor the `sdk` repo to use environment variables more reliably
Load env variables from `const.ts` and make it the single source of truth

Put `const.ts` at the top of the barrel `index.ts` file/entrypoint
