+++ src/typescript/sdk/.env.example
@@ -10,12 +10,6 @@ PUBLISHER_PK="29479e9e5fe47ba9a8af509dd6da1f907510bcf8917bfb19b7079d8c63c0b720"
 # Derived from `PUBLISHER_PK`.
 MODULE_ADDRESS="4bab58978ec1b1bef032eeb285ad47a6a9b997d646c19b598c35f46b26ff9ece"
 
-# Randomize the publisher for ease of use when testing. This facilitates "wiping"
-# the contract state without having to reset the local Aptos node.
-# Note that this will override the `PUBLISHER_PK` and `MODULE_ADDRESS` values if
-# set to "true".
-RANDOMIZE_PUBLISHER="false"
-
 # URL of the Inbox deployment
 # Warning: no / at the end
 INBOX_URL="http://localhost:3000"
