+++ src/typescript/sdk/src/emojicoin_dot_fun/index.ts
@@ -1,4 +1,3 @@
 export * as EmojicoinDotFun from "./emojicoin-dot-fun";
-export * from "./const";
 export * from "./types";
 export * from "./utils";
