+++ src/typescript/sdk/src/emojicoin_dot_fun/utils.ts
@@ -6,7 +6,7 @@ import {
   type UserTransactionResponse,
   MoveString,
 } from "@aptos-labs/ts-sdk";
-import { EMOJICOIN_DOT_FUN_MODULE_NAME, MODULE_ADDRESS } from "./const";
+import { EMOJICOIN_DOT_FUN_MODULE_NAME, MODULE_ADDRESS } from "../const";
 import { type Events, converter, toGenericEvent } from "./events";
 import { type ContractTypes } from "../types";
 import { TYPE_TAGS } from "../utils/type-tags";
