+++ src/typescript/sdk/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./const";
 export * from "./emoji_data";
 export * from "./emojicoin_dot_fun";
 export * from "./markets";
