+++ src/typescript/sdk/src/markets/utils.ts
@@ -9,14 +9,10 @@ import {
   type InputGenerateTransactionOptions,
   parseTypeTag,
 } from "@aptos-labs/ts-sdk";
-import {
-  DEFAULT_REGISTER_MARKET_GAS_OPTIONS,
-  EmojicoinDotFun,
-  COIN_FACTORY_MODULE_NAME,
-  deriveEmojicoinPublisherAddress,
-} from "../emojicoin_dot_fun";
+import { EmojicoinDotFun, deriveEmojicoinPublisherAddress } from "../emojicoin_dot_fun";
 import { type EmojicoinInfo } from "../types/contract";
 import { toConfig } from "../utils/misc";
+import { COIN_FACTORY_MODULE_NAME, DEFAULT_REGISTER_MARKET_GAS_OPTIONS } from "../const";
 
 /**
  * Get the derived market address and TypeTags for the given registry address and symbol bytes.
