+++ src/typescript/sdk/tests/e2e/chat.test.ts
@@ -1,7 +1,8 @@
 import { Account } from "@aptos-labs/ts-sdk";
 import { PostgrestClient } from "@supabase/postgrest-js";
+import { INBOX_URL, ONE_APT } from "../../src/const";
 import { getRegistryAddress, toChatEvent } from "../../src";
-import { EmojicoinDotFun, INBOX_URL, ONE_APT } from "../../src/emojicoin_dot_fun";
+import { EmojicoinDotFun } from "../../src/emojicoin_dot_fun";
 import { sleep } from "../../src/utils";
 import { getTestHelpers } from "../utils";
 import { getEmojicoinMarketAddressAndTypeTags } from "../../src/markets/utils";
